pyproject.toml
[tool.poetry]
name = "oemof_api"
version = "0.1.0"
description = "API to simulate energysystems using energy datapackages from oemof_tabular"
authors = ["hendrik.huyskens <hendrik.huyskens@rl-institut.de>"]
license = "GNU AFFERO GENERAL PUBLIC LICENSE"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.78.0"
aiofiles = "^0.8.0"
Jinja2 = "^3.1.2"
uvicorn = "^0.18.2"
redis = "^4.3.4"
rq = "^1.10.1"
click = "^7.0"
pandas = "1.1.5"
pyomo = "5.7.2"
tableschema = "1.3.0"
datapackage = "1.5.1"
oemoflex = {git = "https://github.com/rl-institut/oemoflex.git", rev = "dev"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
